♻️ refactor `Button` to use `@react-aria/focus` `useFocusRing`
This will _appear_ the same as before, but now the behavior is closer to
correct. Before when a user clicked, we would blur the button to not show
it's focused style. This was wrong for accessiblity because when a button
is clicked, it should retain the focus to allow tabbing. Now we use
`@react-aria/focus` to show the focus style if the user arrived on the
element with the keyboard only.